using-Windsurf/dualmind_orchestrator/requirements.txt
# DualMind Orchestrator Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
openai>=1.0.0
requests>=2.31.0
transformers>=4.21.0
beautifulsoup4>=4.12.0  # Imported as 'bs4' in code

# Data processing and visualization
matplotlib>=3.7.0
pandas>=2.0.0
fpdf>=1.7.2

# Web interface
gradio>=4.0.0
flask>=2.3.0

# Environment and utilities
python-dotenv>=1.0.0

# Optional: For enhanced functionality (will use fallbacks if not available)
# tiktoken>=0.4.0  # For token counting
# pillow>=9.0.0    # For image processing in PDFs

# Note: Some packages may require additional system dependencies
# For transformers: pip install torch (CPU version recommended for demo)
# For matplotlib: Your system may need additional graphics libraries
